+++ internal/provider/helpers/utils.go
@@ -76,6 +76,22 @@ func GetInt64List(result []gjson.Result) types.List {
 	return types.ListValueMust(types.Int64Type, v)
 }
 
+func GetStringSet(result []gjson.Result) types.Set {
+	v := make([]attr.Value, len(result))
+	for r := range result {
+		v[r] = types.StringValue(result[r].String())
+	}
+	return types.SetValueMust(types.StringType, v)
+}
+
+func GetInt64Set(result []gjson.Result) types.Set {
+	v := make([]attr.Value, len(result))
+	for r := range result {
+		v[r] = types.Int64Value(result[r].Int())
+	}
+	return types.SetValueMust(types.Int64Type, v)
+}
+
 func Must[T any](v T, err error) T {
 	if err != nil {
 		panic(err)